ingestion: Fix aws.py job registration
* Moved get_parser_module() to common_lib
* aws.py now uses common_lib.get_parser_module()
* Updated job_definition() call in register() to new parameter list
