+++ ingestion/functions/aws.py
@@ -33,7 +33,8 @@ def get_parser_name_source(source_id, env):
         )
     except RuntimeError:
         return (False, None)
-    return (True, parser) if len(parser.split(".")) == 2 else (True, None)
+    parser_module = common_lib.get_parser_module(parser)  # parsing.folder.subfolder
+    return (True, parser_module) if parser_module.count(".") == 2 else (True, None)
 
 
 def job_definition(
@@ -144,13 +145,14 @@ deregister    Deregister a Batch job definition
             )
             sys.exit(1)
         if parser_name:
-            print(f"Source {args.source_id} will be parsed by parsing.{parser_name}")
-            source_name = parser_name.replace(".", "-")
+            print(f"Source {args.source_id} will be parsed by {parser_name}")
+            source_name = parser_name.replace(".", "-").replace("parsing-", "")
             print(f"Registering job definition for source {source_name}")
             pprint(
                 self.batch_client.register_job_definition(
                     **job_definition(
                         source_name,
+                        args.source_id,
                         args.env,
                         args.cpu,
                         args.memory,
@@ -158,8 +160,8 @@ deregister    Deregister a Batch job definition
                     )
                 )
             )
-            if args.parser == parser_name:
-                print(f"Source {args.source_id} will be parsed by parsing.{parser_name}")
+            if "parsing." + args.parser == parser_name:
+                print(f"Source {args.source_id} will be parsed by {parser_name}")
             else:
                 print(f"Parser {parser_name} for source {args.source_id} in environment {args.env} does not match input {args.parser}")
                 sys.exit(1)
